Exception handling update

+++ bot_logic/admin_logic/admin_handlers.py
@@ -78,15 +78,19 @@ async def neuro_description_added(message: Message, state: FSMContext):
 
 @admin_router.callback_query(F.data.startswith('choose_neuro_type_'), admin_states.CreateNewNeuro.neuro_type)
 async def neuro_type_create_or_choose(callback : CallbackQuery, state: FSMContext):
-    neyro_type_id = re.search(r'\d+', callback.data)
-    await callback.answer(None)
-    if not neyro_type_id:
-        await state.set_state(admin_states.CreateNewNeuro.new_neuro_type)
-        await callback.message.answer('Створимо новий тип. Введіть його назву')
-    else:
-        await state.update_data(neuro_type = int(neyro_type_id.group()))
-        await state.set_state(admin_states.CreateNewNeuro.neuro_video_tutorial)
-        await callback.message.answer('Введіть посилання на відос з ютубу якщо таке є, якщо нема то відправте просто крапку -> .')
+    try:
+        neyro_type_id = re.search(r'\d+', callback.data)
+        await callback.answer(None)
+        if not neyro_type_id:
+            await state.set_state(admin_states.CreateNewNeuro.new_neuro_type)
+            await callback.message.answer('Створимо новий тип. Введіть його назву')
+        else:
+            await state.update_data(neuro_type = int(neyro_type_id.group()))
+            await state.set_state(admin_states.CreateNewNeuro.neuro_video_tutorial)
+            await callback.message.answer('Введіть посилання на відос з ютубу якщо таке є, якщо нема то відправте просто крапку -> .')
+    except Exception as e:
+        await state.clear()
+        await callback.message.answer(text=f'Сталася помилка {e}')
 
 
 @admin_router.message(admin_states.CreateNewNeuro.new_neuro_type)
@@ -135,24 +139,28 @@ async def neuro_is_available_added(message: Message, state: FSMContext):
     message_answer = message.text.strip().lower()
     if message_answer == 'так':
         await state.update_data(is_available = True)
-    elif message_answer == 'ні':
+    else:
         await state.update_data(is_available = False)
-    await create_neuro_with_user_info(state=state)
+    await create_neuro_with_user_info(state=state, message=message)
     await state.clear()
     await message.answer('Нова нейронка записана')
     
 
-async def create_neuro_with_user_info(state: FSMContext):
-    data = await state.get_data()
-    new_neuro = NeuralNetwork(
-    name=data["name"],
-    description=data["description"],
-    neuro_type=data["neuro_type"],
-    neuro_video_tutorial=data["neuro_video_tutorial"],
-    neuro_message_ref=data["neuro_message_ref"],
-    neuro_ref = data["neuro_ref"],
-    is_available=data["is_available"])
-    await rq.write_neuro_to_DB(new_neuro)
+async def create_neuro_with_user_info(state: FSMContext, message: Message):
+    try:
+        data = await state.get_data()
+        new_neuro = NeuralNetwork(
+        name=data["name"],
+        description=data["description"],
+        neuro_type=data["neuro_type"],
+        neuro_video_tutorial=data["neuro_video_tutorial"],
+        neuro_message_ref=data["neuro_message_ref"],
+        neuro_ref = data["neuro_ref"],
+        is_available=data["is_available"])
+        await rq.write_neuro_to_DB(new_neuro)
+    except Exception as e:
+        await state.clear()
+        await message.answer(text=f'Сталася помилка {e}')
 
 #===============================================================
 
@@ -165,13 +173,17 @@ async def choose_neuro_to_edit(message : Message):
 
 @admin_router.callback_query(F.data.startswith("start_update_neuro_"))
 async def start_network_update(callback : CallbackQuery, state : FSMContext):
-    update_network_id = int(re.search(r'\d+', callback.data).group())
-    update_network = await rq.get_neuro_by_id(update_network_id)
-    await state.set_state(admin_states.UpdateNeuro.update_start)
-    await state.update_data(network_id = update_network_id)
-    await callback.answer(None)
-    await callback.message.answer('Виберіть поле, яке хочете змінити', 
-    reply_markup=await admin_keyboards.all_network_info(update_network))
+    try:
+        await callback.answer(None)
+        update_network_id = int(re.search(r'\d+', callback.data).group())
+        update_network = await rq.get_neuro_by_id(update_network_id)
+        await state.set_state(admin_states.UpdateNeuro.update_start)
+        await state.update_data(network_id = update_network_id)
+        await callback.message.answer('Виберіть поле, яке хочете змінити', 
+        reply_markup=await admin_keyboards.all_network_info(update_network))
+    except Exception as e:
+        await state.clear()
+        await callback.message.answer(text=f'Сталася помилка {e}')
 
 #=========================== Оновлення імені
 @admin_router.callback_query(admin_states.UpdateNeuro.update_start, F.data.startswith('update_neuro_name'))
@@ -182,12 +194,19 @@ async def get_new_neuro_name(callback : CallbackQuery, state : FSMContext):
 
 
 @admin_router.message(admin_states.UpdateNeuro.new_name)
-async def set_new_neuro_name(message : Message, state : FSMContext):    
-    data = await state.get_data()
-    network_id = data["network_id"]
-    await rq.update_neuro_name(network_id, message.text)
-    await state.clear()
-    await message.answer("Назву змінено")
+async def set_new_neuro_name(message : Message, state : FSMContext):  
+    try:  
+        data = await state.get_data()
+        network_id = data["network_id"]
+        if await rq.check_neuro_exist(network_id):
+            await rq.update_neuro_name(network_id, message.text)            
+            await message.answer("Назву змінено")
+        else :
+            await message.answer('Схоже, ця нейронка була видалена кимось')
+        await state.clear()
+    except Exception as e:
+        await state.clear()
+        await message.answer(text=f'Сталася помилка {e}')
 #===========================
 
 #=========================== Оновлення опису
@@ -199,12 +218,19 @@ async def get_new_neuro_description(callback : CallbackQuery, state : FSMContext
 
 
 @admin_router.message(admin_states.UpdateNeuro.new_description)
-async def set_new_neuro_description(message : Message, state : FSMContext):    
-    data = await state.get_data()
-    network_id = data["network_id"]
-    await rq.update_neuro_description(network_id, message.text)
-    await state.clear()
-    await message.answer("Опис змінено")
+async def set_new_neuro_description(message : Message, state : FSMContext):   
+    try: 
+        data = await state.get_data()
+        network_id = data["network_id"]
+        if await rq.check_neuro_exist(network_id):
+            await rq.update_neuro_description(network_id, message.text)
+            await message.answer("Опис змінено")
+        else:
+            await message.answer('Схоже, ця нейронка була видалена кимось')
+        await state.clear()
+    except Exception as e:
+        await state.clear()
+        await message.answer(text=f'Сталася помилка {e}')
 #===========================
 
 #=========================== Оновлення типу нейронки
@@ -217,30 +243,45 @@ async def update_neuro_type_start(callback : CallbackQuery, state: FSMContext):
 
 @admin_router.callback_query(admin_states.UpdateNeuro.update_neuro_type, F.data.startswith('choose_neuro_type_'))
 async def update_neuro_type(callback : CallbackQuery, state: FSMContext):
-    neyro_type_id = re.search(r'\d+', callback.data)
-    await callback.answer(None)
-    if not neyro_type_id:
-        await state.set_state(admin_states.UpdateNeuro.new_update_neuro_type)
-        await callback.message.answer('Створимо новий тип. Введіть його назву')
-    else:
-        data = await state.get_data()
-        network_id = data["network_id"]
-        await rq.update_neuro_type(network_id, int(neyro_type_id.group()))
+    try:
+        neuro_type_id = re.search(r'\d+', callback.data)
+        await callback.answer(None)
+        if not neuro_type_id:
+            await state.set_state(admin_states.UpdateNeuro.new_update_neuro_type)
+            await callback.message.answer('Створимо новий тип. Введіть його назву')
+        else:
+            data = await state.get_data()
+            network_id = data["network_id"]
+            if await rq.check_neuro_exist(network_id) and await rq.check_neuro_type_exist(neuro_type_id):
+                await rq.update_neuro_type(network_id, int(neuro_type_id.group()))
+                await callback.message.answer('Тип успішно оновлено')
+            else:
+                await callback.message.answer('Схоже, ця нейронка, або тип, були видалені кимось')
+            await state.clear()
+    except Exception as e:
         await state.clear()
-        await callback.message.answer('Тип успішно оновлено')
+        await callback.message.answer(text=f'Сталася помилка {e}')
 
 
 @admin_router.message(admin_states.UpdateNeuro.new_update_neuro_type)
 async def update_neuro_type_create(message: Message, state: FSMContext):
+    try:
+        new_type = NeuroType(name = message.text)            
+        data = await state.get_data()
+        network_id = data["network_id"]
 
-    new_type = NeuroType(name = message.text)
-        
-    data = await state.get_data()
-    network_id = data["network_id"]
-    new_type_id = await rq.write_neuro_type_to_DB(new_type)
-    await rq.update_neuro_type(network_id, new_type_id)
-    await state.clear()
-    await message.answer('Тип успішно створено і оновлено')
+        if await rq.check_neuro_exist(network_id):
+            new_type_id = await rq.write_neuro_type_to_DB(new_type)
+            await rq.update_neuro_type(network_id, new_type_id)
+            await state.clear()
+            await message.answer('Тип успішно створено і оновлено')            
+        else:
+            await message.answer('Схоже нейронка, якій ви хочете оновити тип, була видалена раніше')
+        await state.clear()
+
+    except Exception as e:
+        await state.clear()
+        await message.answer(text=f'Сталася помилка {e}')
 #===========================
 
 #=========================== Оновлення відео-тутору
@@ -252,12 +293,19 @@ async def get_new_neuro_video_tutorial(callback : CallbackQuery, state : FSMCont
 
 
 @admin_router.message(admin_states.UpdateNeuro.update_neuro_video_tutorial)
-async def set_new_neuro_video_tutorial(message : Message, state : FSMContext):    
-    data = await state.get_data()
-    network_id = data["network_id"]
-    await rq.update_neuro_video_tutorial(network_id, message.text)
-    await state.clear()
-    await message.answer("Відео-тутор оновлено")
+async def set_new_neuro_video_tutorial(message : Message, state : FSMContext):  
+    try:  
+        data = await state.get_data()
+        network_id = data["network_id"]
+        if await rq.check_neuro_exist(network_id):
+            await rq.update_neuro_video_tutorial(network_id, message.text)
+            await message.answer("Відео-тутор оновлено")
+        else:
+            await message.answer('Схоже, ця нейронка була видалена кимось')
+        await state.clear()
+    except Exception as e:
+        await state.clear()
+        await message.answer(text=f'Сталася помилка {e}')
 #===========================
 
 #=========================== Оновлення посилання на повідомлення
@@ -269,12 +317,19 @@ async def get_new_neuro_message_ref(callback : CallbackQuery, state : FSMContext
 
 
 @admin_router.message(admin_states.UpdateNeuro.update_neuro_message_ref)
-async def set_new_neuro_message_ref(message : Message, state : FSMContext):    
-    data = await state.get_data()
-    network_id = data["network_id"]
-    await rq.update_neuro_message_ref(network_id, message.text)
-    await state.clear()
-    await message.answer("Повідомлення оновлено")
+async def set_new_neuro_message_ref(message : Message, state : FSMContext):  
+    try:  
+        data = await state.get_data()
+        network_id = data["network_id"]
+        if await rq.check_neuro_exist(network_id):
+            await rq.update_neuro_message_ref(network_id, message.text)            
+            await message.answer("Повідомлення оновлено")
+        else:
+            await message.answer('Схоже, ця нейронка була видалена кимось')
+        await state.clear()
+    except Exception as e:
+        await state.clear()
+        await message.answer(text=f'Сталася помилка {e}')
 #===========================
 
 #=========================== Оновлення посилання на нейронку
@@ -286,12 +341,19 @@ async def get_new_neuro_ref(callback : CallbackQuery, state : FSMContext):
 
 
 @admin_router.message(admin_states.UpdateNeuro.update_neuro_ref)
-async def set_new_neuro_ref(message : Message, state : FSMContext):    
-    data = await state.get_data()
-    network_id = data["network_id"]
-    await rq.update_neuro_ref(network_id, message.text)
-    await state.clear()
-    await message.answer("Посилання на нейронку оновлено")
+async def set_new_neuro_ref(message : Message, state : FSMContext): 
+    try:   
+        data = await state.get_data()
+        network_id = data["network_id"]
+        if await rq.check_neuro_exist(network_id):
+            await rq.update_neuro_ref(network_id, message.text)
+            await message.answer("Посилання на нейронку оновлено")
+        else:
+            await message.answer('Схоже, ця нейронка була видалена кимось')
+        await state.clear()
+    except Exception as e:
+        await state.clear()
+        await message.answer(text=f'Сталася помилка {e}')
 #===========================
 
 #=========================== Оновлення доступності
@@ -304,17 +366,25 @@ async def get_new_neuro_available(callback : CallbackQuery, state : FSMContext):
 
 @admin_router.message(admin_states.UpdateNeuro.update_is_available)
 async def set_new_neuro_available(message: Message, state: FSMContext):
-    message_answer = message.text.strip().lower()
-    is_available = False
-    if message_answer == 'так':
-        is_available = True
-    elif message_answer == 'ні':
+    try:
+        message_answer = message.text.strip().lower()
         is_available = False
+        if message_answer == 'так':
+            is_available = True
+        elif message_answer == 'ні':
+            is_available = False
 
-    data = await state.get_data()
-    network_id = data["network_id"]
-    await rq.update_neuro_is_available(network_id, is_available)
-    await message.answer('Доступ до читання оновлено')
+        data = await state.get_data()
+        network_id = data["network_id"]
+        if await rq.check_neuro_exist(network_id):
+            await rq.update_neuro_is_available(network_id, is_available)
+            await message.answer('Доступ до читання оновлено')
+        else:
+            await message.answer('Схоже, ця нейронка була видалена кимось')
+        await state.clear()
+    except Exception as e:
+        await state.clear()
+        await message.answer(text=f'Сталася помилка {e}')
 #===========================
 
 #===============================================================
@@ -328,23 +398,39 @@ async def get_neuro_type_for_delete(message : Message):
 
 @admin_router.callback_query(F.data.startswith('delete_neuro_type_'))
 async def confirm_neuro_type_delete(callback : CallbackQuery, state : FSMContext):
-    neuro_type_id = int(re.search(r'\d+', callback.data).group())
-    await callback.answer(None)
-    await state.update_data(delete_neuro_type = neuro_type_id)
-    await state.set_state(admin_states.DelteNeuroType.confirm_neuro_type_delete)
-    await callback.message.answer('Напишіть \'так\', щоб підтвердити видалення, або \'ні\', щоб скасувати')
+    try:
+        await callback.answer(None)
+        neuro_type_id = int(re.search(r'\d+', callback.data).group())
+        if await rq.check_neuro_type_exist(neuro_type_id):
+            await state.update_data(delete_neuro_type = neuro_type_id)
+            await state.set_state(admin_states.DelteNeuroType.confirm_neuro_type_delete)
+            await callback.message.answer('Напишіть \'так\', щоб підтвердити видалення, або \'ні\', щоб скасувати')
+            await callback.message.answer('Схоже, цей тип вже було видалено раніше')
+        else:
+            await callback.message.answer('Схоже, цей тип вже було видалено раніше')
+    except Exception as e:
+        await state.clear()
+        await callback.message.answer(text=f'Сталася помилка {e}')
 
 
 @admin_router.message(admin_states.DelteNeuroType.confirm_neuro_type_delete)
 async def delete_neuro_type(message : Message, state :FSMContext):
-    message_answer = message.text.strip().lower()
-    if message_answer == 'так':
-        data = await state.get_data()
-        await rq.delete_neuro_type(data["delete_neuro_type"])
-        await message.answer('Тип видалено')
-    elif message_answer == 'ні':
-        await message.answer('Видалення типу скасовано')
-    await state.clear()
+    try:
+        message_answer = message.text.strip().lower()
+        if message_answer == 'так':
+            data = await state.get_data()
+            if await rq.check_neuro_type_exist(data["delete_neuro_type"]):
+                await rq.delete_neuro_type(data["delete_neuro_type"])
+                await message.answer('Тип видалено')
+            else:
+                await message.answer('Схоже, цей тип вже видалили раніше')
+        else:
+            await message.answer('Видалення типу скасовано')
+
+        await state.clear()
+    except Exception as e:
+        await state.clear()
+        await message.answer(text=f'Сталася помилка {e}')
     
 
 #===============================================================
@@ -367,14 +453,22 @@ async def confirm_neuro_type_delete(callback : CallbackQuery, state : FSMContext
 
 @admin_router.message(admin_states.DeleteNeuro.confirm_neuro_delete)
 async def delete_neuro_type(message : Message, state :FSMContext):
-    message_answer = message.text.strip().lower()
-    if message_answer == 'так':
-        data = await state.get_data()
-        await rq.delete_neuro(data["delete_neuro"])
-        await message.answer('Нейронку видалено')
-    elif message_answer == 'ні':
-        await message.answer('Видалення нейронки скасовано')
-    await state.clear()
+    try:
+        message_answer = message.text.strip().lower()
+        if message_answer == 'так':
+            data = await state.get_data()
+            if await rq.check_neuro_exist(data["delete_neuro"]):
+                await rq.delete_neuro(data["delete_neuro"])
+                await message.answer('Нейронку видалено')
+            else:
+                await message.answer('Схоже, цю нейронку вже видалили раніше')
+        else:
+            await message.answer('Видалення нейронки скасовано')
+
+        await state.clear()
+    except Exception as e:
+        await state.clear()
+        await message.answer(text=f'Сталася помилка {e}')
     
 
 #===============================================================
